refactor(exception): group exceptions in one file
