+++ src/Drivers/AwsS3.js
@@ -6,8 +6,7 @@
  */
 
 const S3 = require('aws-sdk/clients/s3')
-const FileNotFound = require('../Exceptions/FileNotFound')
-const MethodNotSupported = require('../Exceptions/MethodNotSupported')
+const CE = require('../Exceptions')
 
 class AwsS3 {
   /**
@@ -50,7 +49,7 @@ class AwsS3 {
     return new Promise((resolve, reject) => {
       this.s3.getObject({ Bucket: this.bucket, Key: path }, (err, data) => {
         if (err) {
-          if (err.code === 'NoSuchKey') reject(FileNotFound.file(path))
+          if (err.code === 'NoSuchKey') reject(CE.FileNotFound.file(path))
           return reject(err)
         }
         return resolve(data.Body)
@@ -82,7 +81,7 @@ class AwsS3 {
    * @return {boolean}
    */
   async prepend (path, content) {
-    throw MethodNotSupported.method('prepend', 's3')
+    throw CE.MethodNotSupported.method('prepend', 's3')
   }
 
   /**
@@ -93,7 +92,7 @@ class AwsS3 {
    * @return {boolean}
    */
   async append (path, content) {
-    throw MethodNotSupported.method('append', 's3')
+    throw CE.MethodNotSupported.method('append', 's3')
   }
 
   /**
